# public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Order API
  version: 1.0.0
  description: API для управления заказами
servers:
  - url: http://localhost/api/v1
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Auth-Key
security:
  - ApiKeyAuth: []
paths:
...
  /orders:
    get:
      summary: Получить список заказов
      operationId: getOrders
      parameters:
        - name: done
          in: query
          required: false
          schema:
            type: boolean
          description: Фильтр по завершенным заказам
      responses:
        '200':
          description: Успешное получение списка заказов
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    order_id:
                      type: integer
                      example: 1
                    done:
                      type: boolean
                      example: false
...